MINOR: h2: add a buffer_full flag to the connection
This will be needed to let other streams know that they can't write
because the connection is currently blocked due to a buffer full
condition. Only when some room will free up this flag will be removed,
letting the processing continue. During this time, any stream send
attempt must be postponed with the stream queued to the active list.