MINOR: h2: add rcvd_c and rcvd_s fields to the H2 connections
This is needed so that we know how much data to ack for each stream and
for each connection. We need this because the WINDOW_UPDATE messages
will have to be sent for both and we never know in advance if a message
will fit so we may have to retry sending the update with stored data.
In theory the rcvd_s member, which counts the received data for the
stream, should be stored in the stream, but given that we're sending
ACKs as quickly as possible and in all cases before switching to another
stream, it doesn't make sense to store this into the stream, as it would
needlessly consume memory and further require to pin the stream until the
ACK can be sent.