MEDIUM: h2: split H2_CS_FRAME state in two (headers, payload)
We used to rely on h2c->dsi to determine whether we were waiting for
the frame header or the frame payload, but when dealing with window
update frames it becomes completely bogus since we also have to keep
dsi unmolested when yielding, effectively losing that precious
information.

The real issue is of course that we can resume either to read a frame
header or to read a frame payload, thus we need two different states
for this. Thus now we have H2_CS_FRAME_H and H2_CS_FRAME_P states and
we never need to set h2c->dsi to -1 anymore.