MINOR: h2: add an extra "init" state for the streams
This is needed so that we can yield when waiting for a request buffer
and yet know that we're facing a stream that we've already initialized.
The stream could have stayed in the IDLE state as well and the test
relied on pointer comparisons but that sounds too fragile. This state
must never appear outside of the stream initialization sequence.