WIP/MEDIUM: h2: properly wake up the H2C appctx when sending from the H2S side
When the H2S causes H2C to emit some data, it's mandatory to call the h2c
side's wake() function to try to forward those data otherwise they remain
stuck in the send buffer.

Note that we'll also need the equivalent wakeup between h2c and h2s when
passing data from h2c to h2s (which is not yet implemented).