MEDIUM: h2: replace the HTTP parser with a full parser
Now h2c_frt_make_resp_headers() makes use of the H1 parser instead of
using the HTTP txn. As a benefit, it saves a memory copy by being able
to parse and transcode directly from the h1 output channel to the h2
output buffer.

It was verified that stats and redirect are properly dealt with.