MINOR: h2: add new BODY states to process chunks
When converting H1 to H2, we have to parse chunks and it's much simpler
to have a few more states in the H2 stream to deal with this situation.
This commit adds BSIZE (chunk size), BCHNK (the chunk itself), BCRLF
(the CRLF following the chunk), and TRL0 (start of trailers expected).