WIP/MEDIUM: h2: process chunked H1 response messages
The H1 chunk parser was recopied with some changes as it couldn't be
used as-is (eg: here we're using the buffer's "out" pointer instead of
the "in" pointer).

For now it's still incomplete, because we don't try to put the ES bit on
the last data block (the parser doesn't allow us to know if anything
follows), resulting in a final empty DATA frame only in order to carry
this ES bit. We also don't emit the possible trailers frames, so ES is
unconditionally set on the empty DATA frame. At least now the stats page
loads over H2 even after compression and renders correctly in Firefox.