REORG: h2: rearrange the source files
The code is now rearranged between h2common.{c,h} which only focuses on
the protocol elements and the configuration keywords, and h2ch1.{c,h}
which is specialized on the h2c->h1 gateway. This way it will be easier
to later introduce a clean h2s layer in the middle without having to
break everything again.