WIP/BUG/MEDIUM: h2: mark the h2c appctx as "cant_put" when the buffer is full
Without this we get long pauses during the page load especially when debug
output is in use and slows down progress (or simply by placing usleep(20000)
at the top of debug_hexdump()).

It's not normal that the flags are not set at the moment where the data
fail to be written, so certain things need to be revisited on the write
path here.